Resolve "require confirmation to DESTROY channel"

Closes #224

See merge request team-friendo/signalboost!215

+++ app/services/dispatcher/commands/execute.js
@@ -6,6 +6,7 @@ const deauthorizationRepository = require('../../../db/repositories/deauthorizat
 const phoneNumberService = require('../../../../app/services/registrar/phoneNumber')
 const signal = require('../../signal')
 const logger = require('../logger')
+const { get } = require('lodash')
 const { getAllAdminsExcept } = require('../../../db/repositories/channel')
 const { messagesIn } = require('../strings/messages')
 const { memberTypes } = require('../../../db/repositories/membership')
@@ -50,7 +51,8 @@ const execute = async (executable, dispatchable) => {
     [commands.ACCEPT]: () => maybeAccept(db, channel, sender, language),
     [commands.ADD]: () => maybeAddAdmin(db, sock, channel, sender, payload),
     [commands.DECLINE]: () => decline(db, channel, sender, language),
-    [commands.DESTROY]: () => maybeDestroy(db, sock, channel, sender),
+    [commands.DESTROY]: () => maybeConfirmDestroy(db, sock, channel, sender),
+    [commands.DESTROY_CONFIRM]: () => maybeDestroy(db, sock, channel, sender),
     [commands.HELP]: () => showHelp(db, channel, sender),
     [commands.INFO]: () => showInfo(db, channel, sender),
     [commands.INVITE]: () => maybeInvite(db, channel, sender, payload),
@@ -168,8 +170,22 @@ const decline = async (db, channel, sender, language) => {
 
 // DESTROY
 
+const maybeConfirmDestroy = async (db, sock, channel, sender) => {
+  const cr = messagesIn(sender.language).commandResponses.destroy
+
+  if (!(sender.type === ADMIN)) {
+    return { status: statuses.UNAUTHORIZED, message: cr.notAdmin }
+  }
+  return { status: statuses.SUCCESS, message: cr.confirm }
+}
+
 const maybeDestroy = async (db, sock, channel, sender) => {
   const cr = messagesIn(sender.language).commandResponses.destroy
+
+  if (!(sender.type === ADMIN)) {
+    return { status: statuses.UNAUTHORIZED, message: cr.notAdmin }
+  }
+
   const result = await phoneNumberService.destroy({
     db,
     sock,
@@ -177,7 +193,7 @@ const maybeDestroy = async (db, sock, channel, sender) => {
     sender: sender.phoneNumber,
   })
 
-  if (result.status === statuses.SUCCESS) {
+  if (get(result, 'status') === statuses.SUCCESS) {
     return {
       status: statuses.SUCCESS,
       message: cr.success,
